+++ product-service/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  testEnvironment: "node",
+
+  verbose: false,
+
+  globals: {
+    __DEV__: true,
+  },
+
+  transformIgnorePatterns: ["/node_modules/(?!@)/"],
+
+  rootDir: "./",
+};
