+++ product-service/package.json
@@ -1,9 +1,33 @@
 {
-  "dependencies": {
-    "aws-sdk": "^2.1233.0",
-    "joi": "^17.6.3"
+  "name": "product-service",
+  "version": "1.0.1",
+  "main": "index.js",
+  "license": "MIT",
+  "scripts": {
+    "test": "jest",
+    "swagger": "sls generate-swagger",
+    "deploy": "sls deploy",
+    "remove": "sls remove"
   },
   "devDependencies": {
-    "serverless-webpack": "^5.10.0"
+    "@babel/core": "^7.19.3",
+    "@babel/preset-env": "^7.19.4",
+    "aws-sdk": "^2.1238.0",
+    "aws-sdk-mock": "^5.8.0",
+    "babel-jest": "^29.1.2",
+    "babel-loader": "^8.2.5",
+    "jest": "^29.1.2",
+    "jest-cli": "^29.1.2",
+    "serverless-webpack": "^5.10.0",
+    "webpack": "^5.74.0"
+  },
+  "dependencies": {
+    "csv-parser": "^3.0.0",
+    "serverless-auto-swagger": "^2.10.0",
+    "serverless-offline": "^11.2.3"
+  },
+  "engines": {
+    "node": ">6.9.4",
+    "npm": ">4.2.0"
   }
 }
