+++ product-service/src/handler.js
@@ -1,0 +1,7 @@
+"use strict";
+export {
+  hello,
+  getProductsList,
+  createProduct,
+  getProductsById,
+} from "./handlers";
