+++ product-service/src/handlers/hello.js
@@ -1,0 +1,16 @@
+const hello = async (event) => {
+  console.log("hello Lambda functon ...", event);
+  return {
+    statusCode: 200,
+    body: JSON.stringify(
+      {
+        message: "Go Serverless v1.0! Your function executed successfully!",
+        input: event,
+      },
+      null,
+      2
+    ),
+  };
+};
+
+export { hello };
