+++ product-service/src/handlers/index.js
@@ -1,0 +1,4 @@
+export { hello } from "./hello";
+export { getProductsList } from "./getProductsList";
+export { createProduct } from "./createProduct";
+export { getProductsById } from "./getProductsById";
