+++ product-service/upload-products/aws-scripts.cmd
